refactor: mv providers to own file
