+++ src/connection/index.ts
@@ -8,7 +8,8 @@ import { WalletConnect } from '@web3-react/walletconnect'
 import { SupportedChainId } from 'constants/chains'
 
 import UNISWAP_LOGO_URL from '../assets/svg/logo.svg'
-import { RPC_PROVIDERS, RPC_URLS } from '../constants/networks'
+import { RPC_URLS } from '../constants/networks'
+import { RPC_PROVIDERS } from '../constants/providers'
 
 export enum ConnectionType {
   INJECTED = 'INJECTED',
