+++ src/hooks/useFetchListCallback.ts
@@ -1,7 +1,7 @@
 import { nanoid } from '@reduxjs/toolkit'
 import { TokenList } from '@uniswap/token-lists'
 import { SupportedChainId } from 'constants/chains'
-import { RPC_PROVIDERS } from 'constants/networks'
+import { RPC_PROVIDERS } from 'constants/providers'
 import getTokenList from 'lib/hooks/useTokenList/fetchTokenList'
 import resolveENSContentHash from 'lib/utils/resolveENSContentHash'
 import { useCallback } from 'react'
